+++ src/cli/commands/agent-cdp.ts
@@ -1,7 +1,7 @@
 import { runCmdStreaming } from '../../utils/exec.ts';
 import { AppError } from '../../kernel/errors.ts';
 import { isRemoteBridgeBackend } from './remote-bridge.ts';
-import type { SessionRuntimeHints } from '../../contracts.ts';
+import type { SessionRuntimeHints } from '../../kernel/contracts.ts';
 import type { CliFlags } from '../../utils/cli-flags.ts';
 
 const AGENT_CDP_VERSION = '1.6.0';
