+++ src/cli/commands/connection.ts
@@ -24,7 +24,7 @@ import {
   stopReactDevtoolsCleanup,
 } from './connection-runtime.ts';
 import { writeCommandOutput } from './shared.ts';
-import type { LeaseBackend } from '../../contracts.ts';
+import type { LeaseBackend } from '../../kernel/contracts.ts';
 import type { CliFlags } from '../../utils/cli-flags.ts';
 import type { ClientCommandHandler } from './router-types.ts';
