+++ src/daemon/__tests__/request-router-cost.test.ts
@@ -19,7 +19,7 @@ import { createRequestHandler } from '../request-router.ts';
 import type { DaemonRequest, SessionState } from '../types.ts';
 import { LeaseRegistry } from '../lease-registry.ts';
 import { makeSessionStore } from '../../__tests__/test-utils/store-factory.ts';
-import { daemonCommandRequestSchema } from '../../contracts.ts';
+import { daemonCommandRequestSchema } from '../../kernel/contracts.ts';
 import { emitDiagnostic } from '../../utils/diagnostics.ts';
 
 const mockDispatch = vi.mocked(dispatchCommand);
