+++ src/daemon/lease-context.ts
@@ -1,5 +1,5 @@
 import type { DaemonRequest } from './types.ts';
-import type { LeaseBackend } from '../contracts.ts';
+import type { LeaseBackend } from '../kernel/contracts.ts';
 import type { DeviceLease } from './lease-registry.ts';
 import type { RunnerLogicalLeaseContext } from '../core/runner-lease-context.ts';
 import { stripUndefined } from '../utils/parsing.ts';
