+++ src/utils/cli-flags.ts
@@ -14,7 +14,7 @@ import {
   type ResponseLevel,
   type SessionRuntimeHints,
   type SessionIsolationMode,
-} from '../contracts.ts';
+} from '../kernel/contracts.ts';
 import type { RemoteConfigMetroOptions } from '../remote-config-schema.ts';
 import {
   SCREENSHOT_SPECIFIC_FLAG_DEFINITIONS,
