+++ src/utils/runtime-transport.ts
@@ -1,5 +1,5 @@
 import { URL } from 'node:url';
-import type { SessionRuntimeHints } from '../contracts.ts';
+import type { SessionRuntimeHints } from '../kernel/contracts.ts';
 import { AppError } from '../kernel/errors.ts';
 
 export type ResolvedRuntimeTransport = {
